+++ cluster-operator/src/main/java/io/strimzi/operator/cluster/operator/assembly/CaReconciler.java
@@ -375,38 +375,29 @@ public class CaReconciler {
 
     /**
      * Perform a rolling update of the cluster so that CA certificates get added to their truststores, or expired CA
-     * certificates get removed from their truststores. Note this is only necessary when the CA certificate has changed
-     * due to a new CA key. It is not necessary when the CA certificate is replace while retaining the existing key.
+     * certificates get removed from their truststores. Note this is only necessary when the Cluster CA certificate has changed
+     * due to a new CA key. It is not necessary when the CA certificate is renewed while retaining the existing key.
      */
     Future<Void> rollingUpdateForNewCaKey() {
-        RestartReasons podRollReasons = RestartReasons.empty();
-
         // cluster CA needs to be fully trusted
         // it is coming from a cluster CA key replacement which didn't end successfully (i.e. CO stopped) and we need to continue from there
         if (clusterCa.keyReplaced() || isClusterCaNeedFullTrust) {
-            podRollReasons.add(RestartReason.CLUSTER_CA_CERT_KEY_REPLACED);
-        }
-
-        if (clientsCa.keyReplaced()) {
-            podRollReasons.add(RestartReason.CLIENT_CA_CERT_KEY_REPLACED);
-        }
-
-        if (podRollReasons.shouldRestart()) {
+            String restartReason = RestartReason.CLUSTER_CA_CERT_KEY_REPLACED.getDefaultNote();
             TlsPemIdentity coTlsPemIdentity = new TlsPemIdentity(new PemTrustSet(clusterCa.caCertSecret()), PemAuthIdentity.clusterOperator(coSecret));
             return getZooKeeperReplicas()
-                    .compose(replicas -> maybeRollZookeeper(replicas, podRollReasons, coTlsPemIdentity))
+                    .compose(replicas -> rollZookeeper(replicas, restartReason, coTlsPemIdentity))
                     .compose(i -> getKafkaReplicas())
-                    .compose(nodes -> rollKafkaBrokers(nodes, podRollReasons, coTlsPemIdentity))
-                    .compose(i -> maybeRollDeploymentIfExists(KafkaResources.entityOperatorDeploymentName(reconciliation.name()), podRollReasons))
-                    .compose(i -> maybeRollDeploymentIfExists(KafkaExporterResources.componentName(reconciliation.name()), podRollReasons))
-                    .compose(i -> maybeRollDeploymentIfExists(CruiseControlResources.componentName(reconciliation.name()), podRollReasons));
+                    .compose(nodes -> rollKafkaBrokers(nodes, RestartReasons.of(RestartReason.CLUSTER_CA_CERT_KEY_REPLACED), coTlsPemIdentity))
+                    .compose(i -> rollDeploymentIfExists(KafkaResources.entityOperatorDeploymentName(reconciliation.name()), restartReason))
+                    .compose(i -> rollDeploymentIfExists(KafkaExporterResources.componentName(reconciliation.name()), restartReason))
+                    .compose(i -> rollDeploymentIfExists(CruiseControlResources.componentName(reconciliation.name()), restartReason));
         } else {
             return Future.succeededFuture();
         }
     }
 
     /**
-     * Gather the Kafka related components pods for checking CA key trust and CA certificate usage to sign servers certificate.
+     * Gather the Kafka related components pods for checking Cluster CA key trust and Cluster CA certificate usage to sign servers certificate.
      *
      * Verify that all the pods are already trusting the new CA certificate signed by a new CA key.
      * It checks each pod's CA key generation, compared with the new CA key generation.
@@ -498,33 +489,27 @@ public class CaReconciler {
     }
 
     /**
-     * Checks whether the ZooKeeper cluster needs ot be rolled to trust the new CA private key. ZooKeeper uses only the
-     * Cluster CA and not the Clients CA. So the rolling happens only when Cluster CA private key changed.
+     * Rolls the ZooKeeper cluster to trust the new Cluster CA private key.
      *
      * @param replicas              Current number of ZooKeeper replicas
-     * @param podRestartReasons     List of reasons to restart the pods
+     * @param podRestartReason      Reason to restart the pods
      * @param coTlsPemIdentity      Trust set and identity for TLS client authentication for connecting to ZooKeeper
      *
-     * @return  Future which completes when this step is done either by rolling the ZooKeeper cluster or by deciding
-     *          that no rolling is needed.
+     * @return  Future which completes when the ZooKeeper cluster has been rolled.
      */
-    /* test */ Future<Void> maybeRollZookeeper(int replicas, RestartReasons podRestartReasons, TlsPemIdentity coTlsPemIdentity) {
-        if (podRestartReasons.contains(RestartReason.CLUSTER_CA_CERT_KEY_REPLACED)) {
-            Labels zkSelectorLabels = Labels.EMPTY
-                    .withStrimziKind(reconciliation.kind())
-                    .withStrimziCluster(reconciliation.name())
-                    .withStrimziName(KafkaResources.zookeeperComponentName(reconciliation.name()));
+    /* test */ Future<Void> rollZookeeper(int replicas, String podRestartReason, TlsPemIdentity coTlsPemIdentity) {
+        Labels zkSelectorLabels = Labels.EMPTY
+                .withStrimziKind(reconciliation.kind())
+                .withStrimziCluster(reconciliation.name())
+                .withStrimziName(KafkaResources.zookeeperComponentName(reconciliation.name()));
 
-            Function<Pod, List<String>> rollZkPodAndLogReason = pod -> {
-                List<String> reason = List.of(RestartReason.CLUSTER_CA_CERT_KEY_REPLACED.getDefaultNote());
-                LOGGER.debugCr(reconciliation, "Rolling Pod {} to {}", pod.getMetadata().getName(), reason);
-                return reason;
-            };
-            return new ZooKeeperRoller(podOperator, zookeeperLeaderFinder, operationTimeoutMs)
-                    .maybeRollingUpdate(reconciliation, replicas, zkSelectorLabels, rollZkPodAndLogReason, coTlsPemIdentity);
-        } else {
-            return Future.succeededFuture();
-        }
+        Function<Pod, List<String>> rollZkPodAndLogReason = pod -> {
+            List<String> reason = List.of(podRestartReason);
+            LOGGER.debugCr(reconciliation, "Rolling Pod {} to {}", pod.getMetadata().getName(), reason);
+            return reason;
+        };
+        return new ZooKeeperRoller(podOperator, zookeeperLeaderFinder, operationTimeoutMs)
+                .maybeRollingUpdate(reconciliation, replicas, zkSelectorLabels, rollZkPodAndLogReason, coTlsPemIdentity);
     }
 
     /* test */ Future<Set<NodeRef>> getKafkaReplicas() {
@@ -570,15 +555,6 @@ public class CaReconciler {
                 null,
                 false,
                 eventPublisher);
-    }
-
-    // Entity Operator, Kafka Exporter, and Cruise Control are only rolled when the cluster CA cert key is replaced
-    private Future<Void> maybeRollDeploymentIfExists(String deploymentName, RestartReasons podRollReasons) {
-        if (podRollReasons.contains(RestartReason.CLUSTER_CA_CERT_KEY_REPLACED)) {
-            return rollDeploymentIfExists(deploymentName, RestartReason.CLUSTER_CA_CERT_KEY_REPLACED.getDefaultNote());
-        } else {
-            return Future.succeededFuture();
-        }
     }
 
     /**